Changed int to long in task 9.

+++ src/main/java/com/melnikov/part1/Task9.java
@@ -8,15 +8,15 @@ public class Task9 {
     public static void main(String[] args) {
         String s = JOptionPaneIntegerInputValidator.getValue("Введите число");
 
-        int a = Integer.parseInt(s);
+        long a = Long.parseLong(s);
 
         if (a == 1 || a == 0){
             JOptionPane.showInputDialog("1");
             System.exit(0);
         }
 
-        int fact = 1;
-        int counter = 1;
+        long fact = 1;
+        long counter = 1;
 
         while (counter <= a){
             fact *= counter;
